requirements.txt
faiss==1.5.3
Flask==3.1.2
flask_cors==6.0.1
flask_sqlalchemy==3.1.1
huggingface_hub==0.36.0
numpy==2.3.5
python-dotenv==1.2.1
sentence_transformers==5.1.0
sentence_transformers==5.1.2
